add auth to api

+++ pages/api/registry.js
@@ -23,8 +23,9 @@ export default async (req, res) => {
         let registry = await s3.getObjectJson('hossted-test-reports', 'registry/db.json')
         let machines = registry.data || []
         let url = new URL('https://admin.hossted.com' + req.url)
-        console.log(url.searchParams)
+        console.log(req.headers)
         if (req.method === 'POST') {
+          if (req.headers.authorization == `Basic ${process.env.post_api_key}`) {
             let ip = url.searchParams.get('ip')
             let homepage = url.searchParams.get('url')
             let user = url.searchParams.get('user')
@@ -41,7 +42,12 @@ export default async (req, res) => {
             registry.data = machines
             s3.putObject('hossted-test-reports', 'registry/db.json', JSON.stringify(registry))
             res.send({ message: 'success'})
+          } else {
+            res.send({ message: 'unauthorized'})
+          }
+
         } else if (req.method == 'GET') {
+          if (req.headers.authorization == `Basic ${process.env.read_api_key}`) {
             let ip = url.searchParams.get('ip')
             if (ip) {
                 let machine = getmachine(machines, ip)
@@ -49,18 +55,25 @@ export default async (req, res) => {
             } else {
                 res.status(200).json(machines)
             }
-          } else if (req.method == 'DELETE') {
-            let ip = url.searchParams.get('ip')
-            if (ip) {
-                deleteMachine(machines, ip)
-                registry.data = machines
-                s3.putObject('hossted-test-reports', 'registry/db.json', JSON.stringify(registry))
-                res.send({ message: 'success'})
-            }
-            else {
-                res.send("no ip!")
-            }
+          } else {
+            res.send({ message: 'unauthorized'})
           }
+          } else if (req.method == 'DELETE') {
+            if (req.headers.authorization == `Basic ${process.env.read_api_key}`) {
+              let ip = url.searchParams.get('ip')
+              if (ip) {
+                  deleteMachine(machines, ip)
+                  registry.data = machines
+                  s3.putObject('hossted-test-reports', 'registry/db.json', JSON.stringify(registry))
+                  res.send({ message: 'success'})
+              }
+              else {
+                  res.send("no ip!")
+              }
+          } else {
+            res.send({ message: 'unauthorized'})
+          }
+        }
     } else {
       res.send({ content: domain + ' is not a linnovate or a hossted domain' })
     }
